starting to add Iris theory
I'll have to sort out and reorder the comments and notes on Iris-resources I accumulated so far.
The idea here is to provide an overview of the sources in general and more specifically of
the theory behind it to see if we can boil some of down to a "get an idea what you're working with"
level, while maintaining the link to more detailed explanations.